backend/requirements.txt
# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
aiosqlite==0.19.0  # For testing

# Redis and caching
redis==5.0.1
hiredis==2.2.3

# Background jobs
celery[redis]==5.3.4

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Data validation and serialization
pydantic==2.5.2
pydantic-settings==2.1.0

# HTTP client for external APIs
httpx==0.25.2
aiofiles==23.2.1

# Monitoring and logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.39.2

# Rate limiting
slowapi==0.1.9

# CORS and middleware
starlette==0.27.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
factory-boy==3.3.0
faker==20.1.0

# Code quality
black==23.11.0
isort==5.12.0
mypy==1.7.1
flake8==6.1.0
pre-commit==3.6.0

# Environment management
python-dotenv==1.0.0

# AI/ML integrations (optional)
openai==1.3.7
anthropic==0.7.8

# Utilities
python-slugify==8.0.1
typing-extensions==4.8.0